refactor(console): split the data-source plumbing out of the route files
Two findings left, and both were about a file owning more than one subject.

commands/+page.server.ts drew Overall Code Complexity — a whole-file rule, not
a method one. It hosts the data-source actions because that is where the
builder lives, but ~200 lines of form parsing, quota pre-check, rate limiting
and the rehearsal dial are not about commands. They move to
fetch-def-actions.ts; the route keeps the actions and the fail() calls, which
have to stay there for SvelteKit to infer ActionData. 600 -> 475 lines.

setfetchkey did validation, a demo twin, an owner guard and the seal in one
body, and delfetchkey repeated two of those. keyEntryError returns one message
per line, ownerActor returns the session or the refusal so each caller spends
one branch instead of two, and the demo twins get names. The field->message
map went with it: the UI has only ever rendered a single message, so that
shape was carried and never read.

Verified after: the builder still fetches a live sample and renders its tree,
and sealing a key still returns its last4 and updates the list.
